@@ -1,1 +1,1 @@
-2.4.0
+2.5.0
