@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 RSpec.configure do |config|
-  config.include FactoryGirl::Syntax::Methods
+  config.include FactoryBot::Syntax::Methods
 end
